To compute the average rating of gaming apps, we can use the same approach we used on the previous screen when we computed the average rating of free and non-free apps. In the code example below, we do the following:
Initialize an empty list named games_ratings.
Loop through apps_data[1:], where apps_data is a list of lists that stores our dataset. For each iteration, we do the following:
Assign the rating as a float to a variable named rating.
Assign the genre to a variable named genre and the genre will be saved as a string.
Append the rating value stored in rating to the list games_ratings if the value in genre is equal to the string 'Games'.
Compute the average rating of gaming apps and assign the result to avg_rating_games.
Print avg_rating_games.